Inject a nonce value to the commit headers
Instead of walking authorship or commit dates backwards, inject a nonce
field into the git commit object headers.  Other git commands basically
ignore unrecognized metadata fields, so this is probably harmless.  It has
the nice property of preserving the original author and commit dates.

The idea is shamelessly stolen from Kevin Wallace's Golang gitbrute.

The fieldname of the nonce defaults to 'x-gitbrutec-nonce', but may be
overridden with the '-N fieldname' option to gitbrutec.  The field contents
are a random 64-bit value encoded as a fixed-length base-26 string (A-Z).