If an existing nonce of the same name exists, rewrite it
Don't keep appending new nonce headers if a commit is amended, for
example -- just rewrite the previous one.

This parsing is pretty naive/lazy and assumes (1) our nonce is the very
last header, and (2) that the prior nonce length was exactly what we
expect.  In the cases where we erroneously fail to detect our existing
nonce, we will just add a new field; not a big deal.  In the case where
the prior length was wrong, it was not created with this tool and we may
generate trash.

While here, periodically re-roll the nonce from random and only use
incremental update between re-rolls.  Check for suicide/win condition
much less frequently.  Add comment explaining why overflow check on
base 26 increment is not necessary.